wandb/run-20220809_181703-3roypbj9/files/requirements.txt
absl-py==1.2.0
ale-py==0.7.5
antlr4-python3-runtime==4.9.3
atari-py==0.2.9
attrs==21.4.0
autorom.accept-rom-license==0.4.2
autorom==0.4.2
black==22.6.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.1.0
cryptography==37.0.1
cython==0.29.30
decorator==4.4.2
distlib==0.3.5
dm-control==1.0.4
dm-env==1.5
dm-tree==0.1.7
docker-pycreds==0.4.0
fasteners==0.17.3
filelock==3.7.1
functorch==0.3.0a0+516a8cd
gitdb==4.0.9
gitpython==3.1.27
glfw==2.5.4
google-auth-oauthlib==0.4.6
google-auth==2.9.1
grpcio==1.47.0
gym-notices==0.0.7
gym==0.24.0
hydra-core==1.2.0
hydra-submitit-launcher==1.2.0
identify==2.5.2
idna==3.3
imageio-ffmpeg==0.4.7
imageio==2.20.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
labmaze==1.0.5
lxml==4.9.1
markdown==3.4.1
markupsafe==2.1.1
moviepy==1.0.3
mujoco-py==2.1.2.14
mujoco==2.2.1
mypy-extensions==0.4.3
ninja==1.10.2.3
nodeenv==1.7.0
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.2.2
packaging==21.3
pathspec==0.9.0
pathtools==0.1.2
pillow==9.2.0
pip==22.1.2
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
proglog==0.1.10
promise==2.3
protobuf==3.20.1
psutil==5.9.1
py==1.11.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pygame==2.1.2
pyopengl==3.1.6
pyopenssl==22.0.0
pyparsing==2.4.7
pysocks==1.7.1
pytest-instafail==0.4.2
pytest==7.1.2
pyyaml==6.0
requests-oauthlib==1.3.1
requests==2.28.1
rsa==4.9
scipy==1.8.1
sentry-sdk==1.8.0
setproctitle==1.2.3
setuptools==61.2.0
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
submitit==1.4.5
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.9.0
toml==0.10.2
tomli==2.0.1
torch==1.13.0.dev20220725
torchrl==0.0.1a0+008bebd
torchvision==0.14.0.dev20220725
tqdm==4.64.0
typing-extensions==4.1.1
urllib3==1.26.9
virtualenv==20.16.0
wandb==0.12.21
werkzeug==2.2.0
wheel==0.37.1
zipp==3.8.1